Improve usability by adding expected file format
