+++ apps/client/src/app/pages/portfolio/activities/import-activities-dialog/interfaces/interfaces.ts
@@ -1,5 +1,6 @@
+import { User } from '@ghostfolio/common/interfaces';
+
 export interface ImportActivitiesDialogParams {
-  activities: any[];
   deviceType: string;
-  messages: string[];
+  user: User;
 }
